✨ feat(Controller): validação de `id`

+++ src/controllers/Controller.ts
@@ -1,5 +1,7 @@
 import { ZodSchema } from 'zod';
+import ParamsSchema from '../interfaces/ParamsSchema';
 import GenericService from '../services/Service';
+import ErrorMessage from '../utils/ErrorMessage';
 import ValidationError from '../utils/ValidationError';
 
 export default class Controller<T> {
@@ -16,15 +18,23 @@ export default class Controller<T> {
 
   read = async (): Promise<T[]> => this.service.read();
 
-  readOne = async (id: string): Promise<T | null> =>
-    this.service.readOne(id);
+  readOne = async (id: string): Promise<T | null> => {
+    const { success: idOkay } = ParamsSchema.safeParse(id);
+    if (!idOkay) throw new ValidationError(ErrorMessage.ID_ERROR);
+    return this.service.readOne(id);
+  };
 
   update = async (id: string, data: T): Promise<T | null> => {
+    const { success: idOkay } = ParamsSchema.safeParse(id);
+    if (!idOkay) throw new ValidationError(ErrorMessage.ID_ERROR);
     const { success } = this.validator.safeParse(data);
     if (!success) throw new ValidationError();
     return this.service.update(id, data);
   };
 
-  delete = async (id: string): Promise<T | null> =>
-    this.service.delete(id);
+  delete = async (id: string): Promise<T | null> => {
+    const { success: idOkay } = ParamsSchema.safeParse(id);
+    if (!idOkay) throw new ValidationError(ErrorMessage.ID_ERROR);
+    return this.service.delete(id);
+  };
 }
